+++ index.html
@@ -234,7 +234,6 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="js/functions.js" type="module"></script>
   <script src="js/main.js" type="module"></script>
 </body>
