+++ js/script.js
@@ -1,0 +1,1 @@
+import './util.js';
